+++ pkg/conditions/ready.go
@@ -1,0 +1,25 @@
+package conditions
+
+import (
+	capi "sigs.k8s.io/cluster-api/api/v1alpha3"
+	capiconditions "sigs.k8s.io/cluster-api/util/conditions"
+)
+
+// IsReadyTrue checks if specified object is in Ready condition (if Ready
+// condition is set with status True).
+func IsReadyTrue(object Object) bool {
+	return capiconditions.IsTrue(object, capi.ReadyCondition)
+}
+
+// IsReadyFalse checks if specified object is not in Ready condition (if Ready
+// condition is set with status False).
+func IsReadyFalse(object Object) bool {
+	return capiconditions.IsFalse(object, capi.ReadyCondition)
+}
+
+// IsReadyUnknown checks if it is unknown whether the specified object is in
+// Ready condition or not (if Ready condition is not set, or it is set with
+// status Unknown).
+func IsReadyUnknown(object Object) bool {
+	return capiconditions.IsUnknown(object, capi.ReadyCondition)
+}